+++ backend/canisters/local_user_index/api/src/lib.rs
@@ -67,6 +67,7 @@ pub enum GroupIndexEvent {
     GroupVerifiedChanged(VerifiedChanged),
     CommunityVerifiedChanged(VerifiedChanged),
     CommunityModerationFlagsChanged(ModerationFlagsChanged),
+    GroupModerationFlagsChanged(ModerationFlagsChanged),
     NotifyOfUserDeleted(CanisterId, UserId),
 }